* master: (44 commits)
  Pull `keys()` from `Object`
  List more missing source files
  Make predicates generic
  Refactor `getStyle()` for improved perf
  List missing source files
  Closes #110: Implement width and height property (#115)
  Tests of SIA-R28 to SIA-R40 excluding SIA-R31 (#114)
  Set up project references for examples
  Export SIA-R40
  Don't allow whitespace in ID reference
  Minor refactoring
  SIA-R28 should look for image buttons
  Don't attempt to scrape non-documents
  Lint the things!
  SIA-R21 should only look for explicit roles
  Export some missing functions
  Add `Result` type
  Implement SIA-R40
  Tests of SIA-R19 to SIA-R27 including SIA-R31 (#95)
  Refactor SIA-R10 and SIA-R21
  ...

+++ packages/alfa-unicode/src/get-category.ts
@@ -6,7 +6,7 @@ const { isArray } = Array;
 // Allocate a byte array with enough slots for the entire valid Unicode range.
 // Each slot represents a Unicode code point and its value the category of the
 // code point.
-const categories = new Uint8Array(0x110000);
+const categories = new Uint32Array(0x110000);
 
 for (let i = 0, n = Characters.length; i < n; i++) {
   const { code, category } = Characters[i];
